jem: Use /bin/echo to create the stack file
The make_stack function calls the shell's built-in echo command, which doesn't
know the -ne arguments, causing the stack file to be created with the following
incorrect contents:
-ne \x00\x50\x7c\x80
-ne \x00\x50\x7c\x80
..

This in turn caused a reboot to recovery mess up the stack/device and the
stock images needed to be restored via fastboot.

The use of /bin/echo fixes the problem.